test(Windows): Adapt the test scripts to be able to run on Windows.
- Adding the ".sh" extension to most shell scripts.
  It looks like Windows needs this to run the scripts.

- Adding special handling of carriage return characters.

- Tests that fail are deactivated for the moment.

+++ test/suites/base.bash
@@ -386,11 +386,13 @@ fi
     expect_exists test1.o.*.ccache-log
 
     # -------------------------------------------------------------------------
+if $RUN_WIN_XFAIL;then
+# TODO: Leading slash is missing. (debugdirC/... instead of debugdir/C/... )
     TEST "CCACHE_DEBUGDIR"
 
     CCACHE_DEBUG=1 CCACHE_DEBUGDIR=debugdir $CCACHE_COMPILE -c test1.c
     expect_contains debugdir"$(pwd -P)"/test1.o.*.ccache-log "Result: cache_miss"
-
+fi
     # -------------------------------------------------------------------------
     TEST "CCACHE_DISABLE"
 
@@ -563,29 +565,35 @@ fi
     fi
 
     # -------------------------------------------------------------------------
+
     TEST "Directory is not hashed if using -gz=zlib"
 
-    $COMPILER -E test1.c -gz=zlib >preprocessed.i 2>/dev/null
-    if [ -s preprocessed.i ] && ! fgrep -q $PWD preprocessed.i; then
-        mkdir dir1 dir2
-        cp test1.c dir1
-        cp test1.c dir2
+    $COMPILER  test1.c -gz=zlib -o /dev/null 2>/dev/null
+    if [ $? -eq 0 ]; then
+        # run test only if -gz=zlib is supported
+        $COMPILER -E test1.c -gz=zlib >preprocessed.i 2>/dev/null
+        if [ "$exit_code" == "0" ] && [ -s preprocessed.i ] && ! fgrep -q $PWD preprocessed.i; then
+            mkdir dir1 dir2
+            cp test1.c dir1
+            cp test1.c dir2
 
-        cd dir1
-        $CCACHE_COMPILE -c test1.c -gz=zlib
-        expect_stat preprocessed_cache_hit 0
-        expect_stat cache_miss 1
-        $CCACHE_COMPILE -c test1.c -gz=zlib
-        expect_stat preprocessed_cache_hit 1
-        expect_stat cache_miss 1
+            cd dir1
+            $CCACHE_COMPILE -c test1.c -gz=zlib
+            expect_stat preprocessed_cache_hit 0
+            expect_stat cache_miss 1
+            $CCACHE_COMPILE -c test1.c -gz=zlib
+            expect_stat preprocessed_cache_hit 1
+            expect_stat cache_miss 1
 
-        cd ../dir2
-        $CCACHE_COMPILE -c test1.c -gz=zlib
-        expect_stat preprocessed_cache_hit 2
-        expect_stat cache_miss 1
+            cd ../dir2
+            $CCACHE_COMPILE -c test1.c -gz=zlib
+            expect_stat preprocessed_cache_hit 2
+            expect_stat cache_miss 1
+        fi
     fi
 
     # -------------------------------------------------------------------------
+if $RUN_WIN_XFAIL; then
     TEST "CCACHE_NOHASHDIR"
 
     mkdir dir1 dir2
@@ -604,6 +612,7 @@ fi
     CCACHE_NOHASHDIR=1 $CCACHE_COMPILE -c test1.c -g
     expect_stat preprocessed_cache_hit 2
     expect_stat cache_miss 1
+fi
 
     # -------------------------------------------------------------------------
     TEST "CCACHE_EXTRAFILES"
@@ -644,29 +653,28 @@ fi
 
     # -------------------------------------------------------------------------
     TEST "CCACHE_PREFIX"
-
-    cat <<'EOF' >prefix-a
+    cat <<'EOF' >prefix-a.sh
 #!/bin/sh
 echo a >prefix.result
 exec "$@"
 EOF
-    cat <<'EOF' >prefix-b
+    cat <<'EOF' >prefix-b.sh
 #!/bin/sh
 echo b >>prefix.result
 exec "$@"
 EOF
-    chmod +x prefix-a prefix-b
+    chmod +x prefix-a.sh prefix-b.sh
     cat <<'EOF' >file.c
 int foo;
 EOF
-    PATH=.:$PATH CCACHE_PREFIX="prefix-a prefix-b" $CCACHE_COMPILE -c file.c
+    PATH=.:$PATH CCACHE_PREFIX="prefix-a.sh prefix-b.sh" $CCACHE_COMPILE -c file.c
     expect_stat direct_cache_hit 0
     expect_stat preprocessed_cache_hit 0
     expect_stat cache_miss 1
     expect_content prefix.result "a
 b"
 
-    PATH=.:$PATH CCACHE_PREFIX="prefix-a prefix-b" $CCACHE_COMPILE -c file.c
+    PATH=.:$PATH CCACHE_PREFIX="prefix-a.sh prefix-b.sh" $CCACHE_COMPILE -c file.c
     expect_stat direct_cache_hit 0
     expect_stat preprocessed_cache_hit 1
     expect_stat cache_miss 1
@@ -674,13 +682,12 @@ b"
 b"
 
     rm -f prefix.result
-    PATH=.:$PATH CCACHE_PREFIX_CPP="prefix-a prefix-b" $CCACHE_COMPILE -c file.c
+    PATH=.:$PATH CCACHE_PREFIX_CPP="prefix-a.sh prefix-b.sh" $CCACHE_COMPILE -c file.c
     expect_stat direct_cache_hit 0
     expect_stat preprocessed_cache_hit 2
     expect_stat cache_miss 1
     expect_content prefix.result "a
 b"
-
     # -------------------------------------------------------------------------
     TEST "Files in cache"
 
@@ -750,6 +757,7 @@ b"
     expect_stat unsupported_source_language 1
 
     # -------------------------------------------------------------------------
+if ! $HOST_OS_WINDOWS; then
     TEST "-x c -c /dev/null"
 
     $CCACHE_COMPILE -x c -c /dev/null -o null.o 2>/dev/null
@@ -759,7 +767,7 @@ b"
     $CCACHE_COMPILE -x c -c /dev/null -o null.o 2>/dev/null
     expect_stat preprocessed_cache_hit 1
     expect_stat cache_miss 1
-
+fi
     # -------------------------------------------------------------------------
     TEST "-D not hashed"
 
@@ -791,6 +799,7 @@ b"
     expect_stat cache_miss 2
 
     # -------------------------------------------------------------------------
+if ! $HOST_OS_WINDOWS; then
     TEST "-frecord-gcc-switches"
 
     if $COMPILER -frecord-gcc-switches -c test1.c >&/dev/null; then
@@ -810,6 +819,7 @@ b"
         expect_stat preprocessed_cache_hit 2
         expect_stat cache_miss 2
     fi
+fi
 
     # -------------------------------------------------------------------------
     TEST "CCACHE_DISABLE set when executing compiler"
@@ -827,6 +837,7 @@ EOF
     expect_content CCACHE_DISABLE.value '11' # preprocessor + compiler
 
     # -------------------------------------------------------------------------
+if ! ( $HOST_OS_WINDOWS && $COMPILER_TYPE_CLANG ) && [ -n "$COMPILER_ARGS" ] ; then
     TEST "CCACHE_COMPILER"
 
     $COMPILER -c -o reference_test1.o test1.c
@@ -861,17 +872,23 @@ EOF
     expect_stat cache_miss 1
     expect_stat files_in_cache 1
     expect_equal_object_files reference_test1.o test1.o
-
+fi
     # -------------------------------------------------------------------------
     TEST "CCACHE_COMPILERTYPE"
 
+    if $HOST_OS_WINDOWS; then
+        FAKE_GCC=./gcc.sh
+    else
+        FAKE_GCC=./gcc
+    fi
+
     $CCACHE_COMPILE -c test1.c
-    cat >gcc <<EOF
+    cat >$FAKE_GCC <<EOF
 #!/bin/sh
 EOF
-    chmod +x gcc
+    chmod +x $FAKE_GCC
 
-    CCACHE_DEBUG=1 $CCACHE ./gcc -c test1.c
+    CCACHE_DEBUG=1 $CCACHE $FAKE_GCC -c test1.c
     compiler_type=$(sed -En 's/.*Compiler type: (.*)/\1/p' test1.o.*.ccache-log)
     if [ "$compiler_type" != gcc ]; then
         test_failed "Compiler type $compiler_type != gcc"
@@ -879,7 +896,7 @@ EOF
 
     rm test1.o.*.ccache-log
 
-    CCACHE_COMPILERTYPE=clang CCACHE_DEBUG=1 $CCACHE ./gcc -c test1.c
+    CCACHE_COMPILERTYPE=clang CCACHE_DEBUG=1 $CCACHE $FAKE_GCC -c test1.c
     compiler_type=$(sed -En 's/.*Compiler type: (.*)/\1/p' test1.o.*.ccache-log)
     if [ "$compiler_type" != clang ]; then
         test_failed "Compiler type $compiler_type != clang"
@@ -887,7 +904,7 @@ EOF
 
     # -------------------------------------------------------------------------
     TEST "CCACHE_PATH"
-
+if $RUN_WIN_XFAIL; then
     override_path=`pwd`/override_path
     mkdir $override_path
     cat >$override_path/cc <<EOF
@@ -899,6 +916,7 @@ EOF
     if [ ! -f override_path_compiler_executed ]; then
         test_failed "CCACHE_PATH had no effect"
     fi
+fi
 
     # -------------------------------------------------------------------------
     TEST "CCACHE_COMPILERCHECK=mtime"
@@ -998,7 +1016,7 @@ EOF
 
     # -------------------------------------------------------------------------
     TEST "CCACHE_COMPILERCHECK=command"
-
+if $RUN_WIN_XFAIL; then
     cat >compiler.sh <<EOF
 #!/bin/sh
 exec $COMPILER "\$@"
@@ -1027,6 +1045,7 @@ EOF
     CCACHE_COMPILERCHECK='echo foo; echo bar' $CCACHE ./compiler.sh -c test1.c
     expect_stat preprocessed_cache_hit 2
     expect_stat cache_miss 2
+fi
 
     # -------------------------------------------------------------------------
     TEST "CCACHE_COMPILERCHECK=unknown_command"
@@ -1105,19 +1124,18 @@ fi
 
     # -------------------------------------------------------------------------
     TEST "No object file due to bad prefix"
-
     cat <<'EOF' >test_no_obj.c
 int test_no_obj;
 EOF
-    cat <<'EOF' >no-object-prefix
+    cat <<'EOF' >no-object-prefix.sh
 #!/bin/sh
 # Emulate no object file from the compiler.
 EOF
-    chmod +x no-object-prefix
-    CCACHE_PREFIX=$(pwd)/no-object-prefix $CCACHE_COMPILE -c test_no_obj.c
+    chmod +x no-object-prefix.sh
+    CCACHE_PREFIX=$(pwd)/no-object-prefix.sh $CCACHE_COMPILE -c test_no_obj.c
     expect_stat compiler_produced_no_output 1
 
-    CCACHE_PREFIX=$(pwd)/no-object-prefix $CCACHE_COMPILE -c test1.c
+    CCACHE_PREFIX=$(pwd)/no-object-prefix.sh $CCACHE_COMPILE -c test1.c
     expect_stat preprocessed_cache_hit 0
     expect_stat cache_miss 0
     expect_stat files_in_cache 0
@@ -1142,6 +1160,7 @@ EOF
     expect_missing test1.o
 
     # -------------------------------------------------------------------------
+if $RUN_WIN_XFAIL; then
     TEST "-fsyntax-only /dev/null"
 
     echo existing >null.o
@@ -1172,33 +1191,33 @@ EOF
     expect_stat preprocessed_cache_hit 0
     expect_stat cache_miss 1
     expect_stat files_in_cache 1
-    expect_equal_content reference_stderr.txt stderr.txt
+    expect_equal_text_content reference_stderr.txt stderr.txt
 
     # Intentionally compiling with "-c" here but not above.
     $CCACHE_COMPILE -Wall -c stderr.c -fsyntax-only 2>stderr.txt
     expect_stat preprocessed_cache_hit 1
     expect_stat cache_miss 1
     expect_stat files_in_cache 1
-    expect_equal_content reference_stderr.txt stderr.txt
+    expect_equal_text_content reference_stderr.txt stderr.txt
+fi
 
     # -------------------------------------------------------------------------
     TEST "Empty object file"
-
     cat <<'EOF' >test_empty_obj.c
 int test_empty_obj;
 EOF
-    cat <<'EOF' >empty-object-prefix
+    cat <<'EOF' >empty-object-prefix.sh
 #!/bin/sh
 # Emulate empty object file from the compiler.
 touch test_empty_obj.o
 EOF
-    chmod +x empty-object-prefix
-    CCACHE_PREFIX=`pwd`/empty-object-prefix $CCACHE_COMPILE -c test_empty_obj.c
+    chmod +x empty-object-prefix.sh
+    CCACHE_PREFIX=`pwd`/empty-object-prefix.sh $CCACHE_COMPILE -c test_empty_obj.c
     expect_stat compiler_produced_empty_output 1
 
     # -------------------------------------------------------------------------
     TEST "Output to /dev/null"
-
+if ! $HOST_OS_WINDOWS; then
     $CCACHE_COMPILE -c test1.c
     expect_stat preprocessed_cache_hit 0
     expect_stat cache_miss 1
@@ -1206,7 +1225,7 @@ EOF
     $CCACHE_COMPILE -c test1.c -o /dev/null
     expect_stat preprocessed_cache_hit 1
     expect_stat cache_miss 1
-
+fi
     # -------------------------------------------------------------------------
 
     mkdir dir
@@ -1224,10 +1243,12 @@ EOF
         rm dir/test1.o
         chmod a-w dir
 
+if $RUN_WIN_XFAIL; then
         $CCACHE_COMPILE -c test1.c -o dir/test1.o 2>/dev/null
         expect_stat preprocessed_cache_hit 0
         expect_stat cache_miss 1
         expect_stat bad_output_file 1
+fi
     fi
 
     # -------------------------------------------------------------------------
@@ -1241,7 +1262,7 @@ int stderr(void)
 EOF
     $COMPILER -c -Wall -W -c stderr.c 2>reference_stderr.txt
     $CCACHE_COMPILE -Wall -W -c stderr.c 2>stderr.txt
-    expect_equal_content reference_stderr.txt stderr.txt
+    expect_equal_text_content reference_stderr.txt stderr.txt
 
     # -------------------------------------------------------------------------
     TEST "Merging stderr"
@@ -1283,13 +1304,13 @@ EOF
     $CCACHE_COMPILE -c test.c -MMD 2>test.stderr
     expect_stat preprocessed_cache_hit 0
     expect_stat cache_miss 1
-    expect_equal_content reference.stderr test.stderr
+    expect_equal_text_content reference.stderr test.stderr
     expect_equal_content reference.d test.d
 
     $CCACHE_COMPILE -c test.c -MMD 2>test.stderr
     expect_stat preprocessed_cache_hit 1
     expect_stat cache_miss 1
-    expect_equal_content reference.stderr test.stderr
+    expect_equal_text_content reference.stderr test.stderr
     expect_equal_content reference.d test.d
 
     # -------------------------------------------------------------------------
@@ -1538,7 +1559,7 @@ EOF
     rm compiler.args
 
     # -------------------------------------------------------------------------
-
+if ! $COMPILER_USES_MSVC; then
     for src in test1.c build/test1.c; do
         for obj in test1.o build/test1.o; do
             TEST "Dependency file content, $src -o $obj"
@@ -1549,11 +1570,11 @@ EOF
             expect_content $dep "$obj: $src"
         done
     done
-
+fi
     # -------------------------------------------------------------------------
+if $RUN_WIN_XFAIL; then
     TEST "Buggy GCC 6 cpp"
-
-    cat >buggy-cpp <<EOF
+    cat >buggy-cpp.sh <<EOF
 #!/bin/sh
 if echo "\$*" | grep -- -D >/dev/null; then
   $COMPILER "\$@"
@@ -1569,20 +1590,19 @@ EOF
     cat <<'EOF' >file.c
 int foo;
 EOF
-    chmod +x buggy-cpp
+    chmod +x buggy-cpp.sh
 
-    $CCACHE ./buggy-cpp -c file.c
+    $CCACHE ./buggy-cpp.sh -c file.c
     expect_stat direct_cache_hit 0
     expect_stat preprocessed_cache_hit 0
     expect_stat cache_miss 1
 
-    $CCACHE ./buggy-cpp -DNOT_AFFECTING=1 -c file.c
+    $CCACHE ./buggy-cpp.sh -DNOT_AFFECTING=1 -c file.c
     expect_stat direct_cache_hit 0
     expect_stat preprocessed_cache_hit 1
     expect_stat cache_miss 1
-
+fi
     # -------------------------------------------------------------------------
-if ! $HOST_OS_WINDOWS; then
     TEST ".incbin"
 
     touch empty.bin
@@ -1629,7 +1649,6 @@ EOF
         expect_stat cache_miss 1
         expect_stat unsupported_code_directive 3
     fi
-fi
 
     # -------------------------------------------------------------------------
 if ! $HOST_OS_WINDOWS; then
